% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsychoMatic-package.R
\docType{package}
\name{PsychoMatic-package}
\alias{PsychoMatic}
\alias{PsychoMatic-package}
\title{PsychoMatic: herramientas psicometricas automatizadas}
\description{
\code{PsychoMatic} reune funciones para descriptivos, analisis factorial
exploratorio, analisis factorial confirmatorio e invarianza de medicion en
flujos de trabajo pensados para R y RStudio.
}
\details{
Funciones principales: \code{desc_auto()}, \code{cormat()}, \code{efa_auto()},
\code{export_efa()}, \code{cfa_auto()}, \code{export_cfa()}, \code{inv_align_auto()} y
\code{factorial_invariance_auto()}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gmoncayoj/PsychoMatic}
  \item Report bugs at \url{https://github.com/gmoncayoj/PsychoMatic/issues}
}

}
\author{
\strong{Maintainer}: Jose Gamarra-Moncayo \email{gamarramoncayoj@gmail.com} (\href{https://orcid.org/0000-0002-0781-3616}{ORCID})

}
\keyword{internal}
